Refactoring Map/List manipulation into own BuiltinsRegistry
