+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/builtin/Builtins.java
@@ -1,34 +1,16 @@
 package org.enso.interpreter.runtime.builtin;
 
 import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.TruffleOptions;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.nio.charset.StandardCharsets;
-import java.util.AbstractMap;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
 import org.enso.compiler.Passes;
 import org.enso.compiler.context.CompilerContext;
 import org.enso.compiler.context.FreshNameSupply;
-import org.enso.compiler.core.CompilerError;
 import org.enso.compiler.phase.BuiltinsIrBuilder;
 import org.enso.interpreter.EnsoLanguage;
 import org.enso.interpreter.node.expression.builtin.Any;
 import org.enso.interpreter.node.expression.builtin.Boolean;
 import org.enso.interpreter.node.expression.builtin.Builtin;
-import org.enso.interpreter.node.expression.builtin.BuiltinRootNode;
 import org.enso.interpreter.node.expression.builtin.Nothing;
 import org.enso.interpreter.node.expression.builtin.Polyglot;
 import org.enso.interpreter.node.expression.builtin.debug.Debug;
@@ -50,27 +32,12 @@ import org.enso.interpreter.node.expression.builtin.runtime.Context;
 import org.enso.interpreter.node.expression.builtin.text.Text;
 import org.enso.interpreter.runtime.EnsoContext;
 import org.enso.interpreter.runtime.Module;
-import org.enso.interpreter.runtime.callable.function.Function;
 import org.enso.interpreter.runtime.data.Type;
 import org.enso.interpreter.runtime.scope.ModuleScope;
-import org.enso.interpreter.runtime.util.CachingSupplier;
 import org.enso.pkg.QualifiedName;
 
 /** Container class for static predefined atoms, methods, and their containing scope. */
 public final class Builtins {
-
-  private static List<Constructor<? extends Builtin>> loadedBuiltinConstructors;
-  private static Map<String, LoadedBuiltinMetaMethod> loadedBuiltinMethodsMeta;
-  private static Map<String, LoadedBuiltinMethod> loadedBuiltinMethods;
-
-  static {
-    loadedBuiltinConstructors = readBuiltinTypes();
-    loadedBuiltinMethodsMeta = readBuiltinMethodsMeta();
-    if (TruffleOptions.AOT) {
-      loadedBuiltinMethods = loadBuiltinMethodClassesEarly(loadedBuiltinMethodsMeta);
-    }
-  }
-
   public static final String PACKAGE_NAME = "Builtins";
   public static final String NAMESPACE = "Standard";
   public static final String MODULE_NAME = NAMESPACE + "." + PACKAGE_NAME + ".Main";
@@ -83,9 +50,7 @@ public final class Builtins {
   }
 
   private final EnsoContext context;
-  private final Map<Class<? extends Builtin>, Builtin> builtins;
-  private final Map<String, Map<String, Supplier<LoadedBuiltinMethod>>> builtinMethodNodes;
-  private final Map<String, Builtin> builtinsByName;
+  private final BuiltinsRegistry builtins;
 
   private final Error error;
   private final Module module;
@@ -137,164 +102,43 @@ public final class Builtins {
     module.compileScope(context); // Dummy compilation for an empty module
     ModuleScope.Builder scopeBuilder = module.newScopeBuilder();
 
-    builtins = initializeBuiltinTypes(loadedBuiltinConstructors, language, scopeBuilder);
-    builtinsByName =
-        builtins.values().stream()
-            .collect(
-                Collectors.toMap(
-                    v -> v.getType().getName(), java.util.function.Function.identity()));
-    builtinMethodNodes = registerBuiltinMethodsLazily(scopeBuilder, language);
+    builtins = new BuiltinsRegistry(language, scopeBuilder);
+
     ordering = getBuiltinType(Ordering.class);
     comparable = getBuiltinType(Comparable.class);
     defaultComparator = getBuiltinType(DefaultComparator.class);
     bool = this.getBuiltinType(Boolean.class);
     contexts = this.getBuiltinType(Context.class);
 
-    any = builtins.get(Any.class);
-    nothing = builtins.get(Nothing.class);
-    function = builtins.get(org.enso.interpreter.node.expression.builtin.function.Function.class);
-    polyglot = builtins.get(Polyglot.class);
-    text = builtins.get(Text.class);
-    array = builtins.get(Array.class);
-    vector = builtins.get(Vector.class);
-    dictionary = builtins.get(org.enso.interpreter.node.expression.builtin.Dictionary.class);
-    dataflowError = builtins.get(org.enso.interpreter.node.expression.builtin.Error.class);
-    ref = builtins.get(Ref.class);
-    managedResource = builtins.get(ManagedResource.class);
-    debug = builtins.get(Debug.class);
+    any = getBuiltinType(Any.class);
+    nothing = getBuiltinType(Nothing.class);
+    function = getBuiltinType(org.enso.interpreter.node.expression.builtin.function.Function.class);
+    polyglot = getBuiltinType(Polyglot.class);
+    text = getBuiltinType(Text.class);
+    array = getBuiltinType(Array.class);
+    vector = getBuiltinType(Vector.class);
+    dictionary = getBuiltinType(org.enso.interpreter.node.expression.builtin.Dictionary.class);
+    dataflowError = getBuiltinType(org.enso.interpreter.node.expression.builtin.Error.class);
+    ref = getBuiltinType(Ref.class);
+    managedResource = getBuiltinType(ManagedResource.class);
+    debug = getBuiltinType(Debug.class);
     projectDescription = getBuiltinType(ProjectDescription.class);
-    file = builtins.get(File.class);
-    date = builtins.get(org.enso.interpreter.node.expression.builtin.date.Date.class);
-    dateTime = builtins.get(org.enso.interpreter.node.expression.builtin.date.DateTime.class);
-    duration = builtins.get(org.enso.interpreter.node.expression.builtin.date.Duration.class);
-    timeOfDay = builtins.get(org.enso.interpreter.node.expression.builtin.date.TimeOfDay.class);
-    timeZone = builtins.get(org.enso.interpreter.node.expression.builtin.date.TimeZone.class);
-    warning = builtins.get(Warning.class);
+    file = getBuiltinType(File.class);
+    date = getBuiltinType(org.enso.interpreter.node.expression.builtin.date.Date.class);
+    dateTime = getBuiltinType(org.enso.interpreter.node.expression.builtin.date.DateTime.class);
+    duration = getBuiltinType(org.enso.interpreter.node.expression.builtin.date.Duration.class);
+    timeOfDay = getBuiltinType(org.enso.interpreter.node.expression.builtin.date.TimeOfDay.class);
+    timeZone = getBuiltinType(org.enso.interpreter.node.expression.builtin.date.TimeZone.class);
+    warning = getBuiltinType(Warning.class);
     noWrap = getBuiltinType(NoWrap.class);
     problemBehavior = getBuiltinType(ProblemBehavior.class);
     additionalWarnings = getBuiltinType(AdditionalWarnings.class);
-    instrumentor = builtins.get(org.enso.interpreter.node.expression.builtin.Instrumentor.class);
+    instrumentor = getBuiltinType(org.enso.interpreter.node.expression.builtin.Instrumentor.class);
 
     error = new Error(this, context);
     system = new System(this);
     number = new Number(this);
     scope = scopeBuilder.build();
-  }
-
-  private static Map<String, LoadedBuiltinMethod> loadBuiltinMethodClassesEarly(
-      Map<String, LoadedBuiltinMetaMethod> map) {
-    Map<String, LoadedBuiltinMethod> methods = new HashMap<>();
-    map.forEach(
-        (key, value) -> {
-          methods.put(key, value.toMethod());
-        });
-    return methods;
-  }
-
-  /**
-   * Registers builtin methods with their corresponding Atom Constructor's owners. That way
-   * "special" builtin types have builtin methods in the scope without requiring everyone to always
-   * import full stdlib.
-   *
-   * @param scope Builtins scope
-   * @param language The language the resulting function nodes should be associated with
-   */
-  private void registerBuiltinMethods(ModuleScope.Builder scope, EnsoLanguage language) {
-    for (Builtin builtin : builtins.values()) {
-      var type = builtin.getType();
-      Map<String, Supplier<LoadedBuiltinMethod>> methods = builtinMethodNodes.get(type.getName());
-      if (methods != null) {
-        // Register a builtin method iff it is marked as auto-register.
-        // Methods can only register under a type or, if we deal with a static method, it's
-        // eigen-type.
-        // Such builtins are available on certain types without importing the whole stdlib, e.g. Any
-        // or Number.
-        methods.forEach(
-            (key, value) -> {
-              LoadedBuiltinMethod meth = value.get();
-              Type tpe =
-                  meth.isAutoRegister ? (!meth.isStatic() ? type : type.getEigentype()) : null;
-              if (tpe != null) {
-                Optional<BuiltinFunction> fun = meth.toFunction(language, false);
-                fun.ifPresent(f -> scope.registerMethod(tpe, key, f.getFunction()));
-              }
-            });
-      }
-    }
-  }
-
-  /**
-   * Register builtin methods and initialize them lazily in the provided scope. This method differs
-   * from `registerBuiltinMethods` where all methods are initialized by the time they are
-   * registered..
-   *
-   * @param scope Builtins scope
-   * @param language The language the resulting function nodes should be associated with
-   * @return map from types to builtin methods
-   */
-  private Map<String, Map<String, Supplier<LoadedBuiltinMethod>>> registerBuiltinMethodsLazily(
-      ModuleScope.Builder scope, EnsoLanguage language) {
-    Map<String, Map<String, Supplier<LoadedBuiltinMethod>>> builtinMethodNodes = new HashMap<>();
-    Map<String, Map<String, LoadedBuiltinMetaMethod>> builtinMetaMethods = new HashMap<>();
-    loadedBuiltinMethodsMeta.forEach(
-        (fullName, meta) -> {
-          String[] builtinName = fullName.split("\\.");
-          if (builtinName.length != 2) {
-            throw new CompilerError("Invalid builtin metadata for " + fullName);
-          }
-          String builtinMethodOwner = builtinName[0];
-          String builtinMethodName = builtinName[1];
-          var constr = scope.getType(builtinMethodOwner, true);
-          if (constr != null) {
-            Map<String, Supplier<LoadedBuiltinMethod>> atomNodes =
-                getOrUpdate(builtinMethodNodes, constr.getName());
-            atomNodes.put(builtinMethodName, CachingSupplier.wrap(() -> meta.toMethod()));
-
-            Map<String, LoadedBuiltinMetaMethod> atomNodesMeta =
-                getOrUpdate(builtinMetaMethods, constr.getName());
-            atomNodesMeta.put(builtinMethodName, meta);
-          } else {
-            Map<String, Supplier<LoadedBuiltinMethod>> atomNodes =
-                getOrUpdate(builtinMethodNodes, builtinMethodOwner);
-            atomNodes.put(builtinMethodName, CachingSupplier.wrap(() -> meta.toMethod()));
-
-            Map<String, LoadedBuiltinMetaMethod> atomNodesMeta =
-                getOrUpdate(builtinMetaMethods, builtinMethodOwner);
-            atomNodesMeta.put(builtinMethodName, meta);
-          }
-        });
-
-    for (Builtin builtin : builtins.values()) {
-      var type = builtin.getType();
-      Map<String, LoadedBuiltinMetaMethod> methods = builtinMetaMethods.get(type.getName());
-      if (methods != null) {
-        // Register a builtin method iff it is marked as auto-register.
-        // Methods can only register under a type or, if we deal with a static method, it's
-        // eigen-type.
-        // Such builtins are available on certain types without importing the whole stdlib, e.g. Any
-        // or Number.
-        methods.forEach(
-            (key, value) -> {
-              Type tpe =
-                  value.isAutoRegister() ? (!value.isStatic() ? type : type.getEigentype()) : null;
-              if (tpe != null) {
-                Supplier<Function> supplier =
-                    () -> value.toMethod().toFunction(language, false).get().getFunction();
-                scope.registerMethod(tpe, key, supplier);
-              }
-            });
-      }
-    }
-    return builtinMethodNodes;
-  }
-
-  private <T> Map<String, T> getOrUpdate(Map<String, Map<String, T>> map, String key) {
-    Map<String, T> entry = map.get(key);
-    if (entry == null) {
-      entry = new HashMap<>();
-      map.put(key, entry);
-    }
-    return entry;
   }
 
   /**
@@ -330,123 +174,6 @@ public final class Builtins {
   }
 
   /**
-   * Returns a list of supported builtins.
-   *
-   * <p>Builtin types are marked via @BuiltinType annotation. The metadata file represents a single
-   * builtin type per row. The format of the row is as follows: <Enso name of the builtin
-   * type>:<Name of the class representing it>:[<field1>,<field2>,...] where the last column gives a
-   * list of optional type's fields.
-   */
-  private static List<Constructor<? extends Builtin>> readBuiltinTypes() {
-    ClassLoader classLoader = Builtins.class.getClassLoader();
-    List<String> lines;
-    final String NODE_PKG = "org.enso.interpreter.node.expression.builtin";
-    final String META_PATH =
-        "META-INF" + "/" + NODE_PKG.replace('.', '/') + "/BuiltinTypes.metadata";
-    try (InputStream resource = classLoader.getResourceAsStream(META_PATH)) {
-      lines =
-          new BufferedReader(new InputStreamReader(resource, StandardCharsets.UTF_8))
-              .lines()
-              .collect(Collectors.toList());
-    } catch (Exception ioe) {
-      lines = new ArrayList<>();
-      ioe.printStackTrace();
-    }
-
-    return lines.stream()
-        .map(
-            line -> {
-              String[] builtinMeta = line.split(":");
-              if (builtinMeta.length < 2 || builtinMeta.length > 4) {
-                java.lang.System.out.println(Arrays.toString(builtinMeta));
-                throw new CompilerError("Invalid builtin metadata in: " + line);
-              }
-              try {
-                @SuppressWarnings("unchecked")
-                Class<? extends Builtin> clazz =
-                    (Class<? extends Builtin>) Class.forName(builtinMeta[1]);
-
-                // Note: Don't create a new instance of the builtin at this point
-                // because that will be too much for the inliner and won't get
-                // constant folded.
-                return clazz.getConstructor();
-              } catch (ClassNotFoundException | NoSuchMethodException e) {
-                e.printStackTrace();
-                throw new CompilerError("Invalid builtin type entry: " + builtinMeta[1]);
-              }
-            })
-        .collect(Collectors.toList());
-  }
-
-  /** Initialize builting types in the context of the given language and module scope */
-  private Map<Class<? extends Builtin>, Builtin> initializeBuiltinTypes(
-      List<Constructor<? extends Builtin>> constrs,
-      EnsoLanguage language,
-      ModuleScope.Builder scope) {
-    Map<Class<? extends Builtin>, Builtin> builtins = new HashMap<>();
-
-    for (var constr : constrs) {
-      try {
-        Builtin builtin = constr.newInstance();
-        builtins.put(builtin.getClass(), builtin);
-      } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
-        throw new CompilerError("Invalid builtin type entry: " + constr, e);
-      }
-    }
-    for (var b : builtins.values()) {
-      b.initialize(language, scope, builtins);
-    }
-    return builtins;
-  }
-
-  /**
-   * Loads a Map of builtin methods.
-   *
-   * <p>Builtin methods are marked via @BuiltinMethod annotation. THe metadata file represents a
-   * single builtin method per row. The format of the row is as follows: <Fully qualified name of
-   * the builtin method>:<Class name of the builtin method representing it>
-   *
-   * @return A map of builtin method nodes per builtin type name
-   */
-  private static Map<String, LoadedBuiltinMetaMethod> readBuiltinMethodsMeta() {
-    ClassLoader classLoader = Builtins.class.getClassLoader();
-    List<String> lines;
-
-    final String NODE_PKG = "org.enso.interpreter.node.expression.builtin";
-    final String META_PATH =
-        "META-INF" + "/" + NODE_PKG.replace('.', '/') + "/BuiltinMethods.metadata";
-    try (InputStream resource = classLoader.getResourceAsStream(META_PATH)) {
-      lines =
-          new BufferedReader(new InputStreamReader(resource, StandardCharsets.UTF_8))
-              .lines()
-              .collect(Collectors.toList());
-    } catch (Exception ioe) {
-      lines = new ArrayList<>();
-      ioe.printStackTrace();
-    }
-
-    return lines.stream()
-        .map(
-            line -> {
-              String[] builtinMeta = line.split(":");
-              if (builtinMeta.length != 4) {
-                throw new CompilerError("Invalid builtin metadata in: " + line);
-              }
-              String[] builtinName = builtinMeta[0].split("\\.");
-              if (builtinName.length != 2) {
-                throw new CompilerError("Invalid builtin metadata in : " + line);
-              }
-              boolean isStatic = java.lang.Boolean.valueOf(builtinMeta[2]);
-              boolean isAutoRegister = java.lang.Boolean.valueOf(builtinMeta[3]);
-
-              return new AbstractMap.SimpleEntry<>(
-                  builtinMeta[0],
-                  new LoadedBuiltinMetaMethod(builtinMeta[1], isStatic, isAutoRegister));
-            })
-        .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-  }
-
-  /**
    * Returns a builtin method for the provided Atom Constructor and the name, if it exists.
    *
    * @param type Atom Constructor owner of the function
@@ -457,20 +184,7 @@ public final class Builtins {
    */
   public Optional<BuiltinFunction> getBuiltinFunction(
       String type, String methodName, EnsoLanguage language, boolean isStaticInstance) {
-    // TODO: move away from String mapping once Builtins is gone
-    Map<String, Supplier<LoadedBuiltinMethod>> atomNodes = builtinMethodNodes.get(type);
-    if (atomNodes == null) {
-      return Optional.empty();
-    }
-    var supply = atomNodes.get(methodName);
-    if (supply == null) {
-      return Optional.empty();
-    }
-    LoadedBuiltinMethod builtin = supply.get();
-    if (builtin == null) {
-      return Optional.empty();
-    }
-    return builtin.toFunction(language, isStaticInstance);
+    return builtins.getBuiltinFunction(type, methodName, language, isStaticInstance);
   }
 
   public Optional<BuiltinFunction> getBuiltinFunction(
@@ -479,22 +193,15 @@ public final class Builtins {
   }
 
   public <T extends Builtin> T getBuiltinType(Class<T> clazz) {
-    @SuppressWarnings("unchecked")
-    T t = (T) builtins.get(clazz);
-    return t;
+    return builtins.getBuiltinType(clazz);
   }
 
   public Builtin getByRepresentationType(Class<?> clazz) {
-    for (var b : builtins.values()) {
-      if (b.isRepresentedBy(clazz)) {
-        return b;
-      }
-    }
-    return null;
+    return builtins.getByRepresentationType(clazz);
   }
 
   public Builtin getBuiltinType(String name) {
-    return builtinsByName.get(name);
+    return builtins.getBuiltinType(name);
   }
 
   /**
@@ -759,57 +466,5 @@ public final class Builtins {
 
   public EnsoLanguage getLanguage() {
     return context.getLanguage();
-  }
-
-  private static class LoadedBuiltinMetaMethod {
-
-    private LoadedBuiltinMethod method;
-    private final String className;
-    private final boolean staticMethod;
-    private final boolean autoRegister;
-
-    private LoadedBuiltinMetaMethod(String className, boolean staticMethod, boolean autoRegister) {
-      this.className = className;
-      this.staticMethod = staticMethod;
-      this.autoRegister = autoRegister;
-      this.method = null;
-    }
-
-    boolean isStatic() {
-      return staticMethod;
-    }
-
-    boolean isAutoRegister() {
-      return autoRegister;
-    }
-
-    LoadedBuiltinMethod toMethod() {
-      if (method == null) {
-        try {
-          @SuppressWarnings("unchecked")
-          Class<BuiltinRootNode> clazz = (Class<BuiltinRootNode>) Class.forName(className);
-          Method meth = clazz.getMethod("makeFunction", EnsoLanguage.class, boolean.class);
-          method = new LoadedBuiltinMethod(meth, staticMethod, autoRegister);
-        } catch (ClassNotFoundException | NoSuchMethodException e) {
-          throw new CompilerError("Invalid builtin method " + className, e);
-        }
-      }
-      return method;
-    }
-  }
-
-  private record LoadedBuiltinMethod(Method meth, boolean isStatic, boolean isAutoRegister) {
-    Optional<BuiltinFunction> toFunction(EnsoLanguage language, boolean isStaticInstance) {
-      try {
-        var f = (Function) meth.invoke(null, language, isStaticInstance);
-        if (f != null) {
-          var bf = new BuiltinFunction(f, isAutoRegister);
-          return Optional.of(bf);
-        }
-      } catch (Exception e) {
-        e.printStackTrace();
-      }
-      return Optional.empty();
-    }
   }
 }
